pyproject.toml
[tool.poetry]
name = "invest_blog_react"
version = "0.1.0"
description = ""
authors = ["balancy <balancy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
python-dotenv = "^0.19.0"
django-cors-headers = "^3.8.0"
django-filter = "^2.4.0"
djangorestframework-simplejwt = "^4.8.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
mixer = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
